Move some things around, isolate them better, make them easier to find, and write a test
